app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pdfplumber
python-multipart
python-docx